Stay consistent with react-leaflet API
