+++ src/GrommetMarker.jsx
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import styled, { ThemeContext } from 'styled-components'
+import { ThemeContext } from 'styled-components'
 import ReactDOMServer from 'react-dom/server';
 import {
   createElementObject,
@@ -7,60 +7,7 @@ import {
   extendContext,
 } from '@react-leaflet/core';
 import L from 'leaflet';
-import { Box } from 'grommet';
-import { StatusCriticalSmall, StatusGoodSmall, StatusWarningSmall, StatusUnknownSmall } from 'grommet-icons';
-import { adjustPadStyle } from './utils/styles';
-
-const StyledBox = styled(Box)`
-  // to create the pin shape
-  border-bottom-right-radius: 0;
-  transform: rotate(45deg);
-  &:hover {
-    // TO DO revisit "hover" color with designers. Use of "background-layer-overlay" feels strange semantically
-    border: ${props => `${props.theme?.global?.borderSize?.[props.border.size] || props.border.size} solid ${props.theme?.global?.colors?.['background-layer-overlay']}`};
-  }
-`;
-
-const statuses = {
-  critical: {
-    color: 'status-critical',
-    icon: StatusCriticalSmall,
-    borderSize: '3px',
-  },
-  warning: {
-    color: 'status-warning',
-    icon: StatusWarningSmall,
-  },
-  good: {
-    color: 'status-ok',
-    icon: StatusGoodSmall,
-  },
-  unknown: {
-    color: 'status-unknown',
-    icon: StatusUnknownSmall
-  }
-}
-
-const StatusIcon = ({ status, theme }) => {
-  const border = { color: statuses[status].color, size: statuses[status].borderSize || 'small'};
-  const StatusIcon = statuses[status].icon;
-
-  return (
-    <StyledBox 
-      background="background-front"
-      border={border}
-      elevation="medium"
-      flex={false}
-      pad={adjustPadStyle(theme.global.edgeSize.xsmall, theme.global.borderSize[border.size] || border.size)}
-      round
-      theme={theme}
-      width={{min: 'fit-content'}}
-    >
-      {/* style needed to offset for rotation of location pin so triangle is vertical still */}
-      <StatusIcon color={statuses[status].color} size="small" style={status === 'warning' ? { transform: 'rotate(-45deg)' } : undefined} />
-    </StyledBox>
-  );
-}
+import { Pin } from './markers';
 
 const createGrommetMarker = (
   { position, title, alt, icon: iconProp, status },
@@ -70,7 +17,10 @@ const createGrommetMarker = (
   const icon = L.divIcon({
     // 'grommet-marker' class prevents leaflet default divIcon styles
     className: 'grommet-marker',
-    html: ReactDOMServer.renderToString(iconProp || <StatusIcon status={status} theme={theme} />),
+    html: ReactDOMServer.renderToString(React.cloneElement(iconProp, {
+      // necessary to pass theme prop for StyledBox in Pi
+      theme: theme
+    }) || <Pin status="unknown" theme={theme} />),
   });
   const options = { title, alt, icon };
   const marker = new L.Marker(position, options);
